requirements.txt
asgiref==3.11.1
attrs==26.1.0
Django==6.0.4
django-dotenv==1.4.2
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.15
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
